requirements.txt
tqdm
nltk
pillow
docx2txt
num2words